# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atlas Backend API
  version: 0.1.0
  description: >
    OpenAPI documentation for the Atlas backend API. This spec covers the
    auth, users, voice, drafts, analytics, alerts, images, research, and
    trending routes exposed by the Express service.
servers:
  - url: /
    description: Current server
...
security:
  - bearerAuth: []
paths:
...
  /api/users/profile:
    get:
      tags:
        - Users
      summary: Get the authenticated user's profile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Users
      summary: Update the authenticated user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/users/team:
    get:
      tags:
        - Users
      summary: List all analysts for manager/admin views
      responses:
        '200':
          description: Team members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamUsersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/users/push-top-profiles:
    post:
      tags:
        - Users
      summary: Push top-performing voice dimensions to inactive analysts
      responses:
        '200':
          description: Action completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/users/send-nudge:
    post:
      tags:
        - Users
      summary: Send inactive analyst nudges
      responses:
        '200':
          description: Action completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/users/push-style:
    post:
      tags:
        - Users
      summary: Push a manager-selected style to all analysts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushStyleRequest'
      responses:
        '200':
          description: Action completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResultResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/trending/scan:
    post:
      tags:
        - Trending
      summary: Scan for trending topics based on subscriptions
      responses:
        '200':
          description: Alerts created from trending items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/trending/topics:
    get:
      tags:
        - Trending
      summary: List cached trending topics for the current user
      responses:
        '200':
          description: Trending topics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendingTopicsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadGateway:
      description: Upstream AI/provider request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required:
        - ok
        - error
        - timestamp
      properties:
        ok:
          type: boolean
          enum:
            - false
        error:
          type: string
        details:
          nullable: true
          description: Additional validation or provider-specific detail payload.
        timestamp:
          type: string
          format: date-time
    SuccessEnvelope:
      type: object
      required:
        - ok
        - timestamp
      properties:
        ok:
          type: boolean
          enum:
            - true
        timestamp:
          type: string
          format: date-time
...
    ActionResultResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - message
                - affected
              properties:
                message:
                  type: string
                affected:
                  type: integer
...
    Role:
      type: string
      enum:
        - ANALYST
        - MANAGER
        - ADMIN
    OnboardingTrack:
      type: string
      enum:
        - TRACK_A
        - TRACK_B
...
    VoiceMaturity:
      type: string
      enum:
        - BEGINNER
        - INTERMEDIATE
        - ADVANCED
...
    VoiceProfile:
      type: object
      required:
        - id
        - userId
        - humor
        - formality
        - brevity
        - contrarianTone
        - maturity
        - tweetsAnalyzed
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        userId:
          type: string
        humor:
          type: integer
          minimum: 0
          maximum: 100
        formality:
          type: integer
          minimum: 0
          maximum: 100
        brevity:
          type: integer
          minimum: 0
          maximum: 100
        contrarianTone:
          type: integer
          minimum: 0
          maximum: 100
        maturity:
          $ref: '#/components/schemas/VoiceMaturity'
        tweetsAnalyzed:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required:
        - id
        - handle
        - role
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        supabaseId:
          type: string
          nullable: true
        handle:
          type: string
        email:
          type: string
          format: email
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/Role'
        onboardingTrack:
          allOf:
            - $ref: '#/components/schemas/OnboardingTrack'
          nullable: true
        telegramChatId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        voiceProfile:
          allOf:
            - $ref: '#/components/schemas/VoiceProfile'
          nullable: true
    TeamCount:
      type: object
      required:
        - tweetDrafts
        - sessions
      properties:
        tweetDrafts:
          type: integer
        sessions:
          type: integer
...
    TeamUser:
      allOf:
        - $ref: '#/components/schemas/UserProfile'
        - type: object
          required:
            - _count
          properties:
            _count:
              $ref: '#/components/schemas/TeamCount'
...
    Alert:
      type: object
      required:
        - id
        - type
        - title
        - createdAt
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        context:
          type: string
          nullable: true
        draftReply:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        sentiment:
          type: string
          nullable: true
        relevance:
          type: number
          format: float
          nullable: true
        userId:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    TrendingTopic:
      type: object
      required:
        - id
        - topic
        - headline
        - context
        - relevance
      properties:
        id:
          type: string
        topic:
          type: string
        headline:
          type: string
        context:
          type: string
        sourceUrl:
          type: string
          nullable: true
        sentiment:
          type: string
          nullable: true
        relevance:
          type: number
          format: float
...
    UpdateUserProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
          description: Accepted by the validator but not currently persisted by the route handler.
        avatarUrl:
          type: string
    PushStyleRequest:
      type: object
      properties:
        blendId:
          type: string
...
    UserProfileResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - user
              properties:
                user:
                  $ref: '#/components/schemas/UserProfile'
    TeamUsersResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - team
              properties:
                team:
                  type: array
                  items:
                    $ref: '#/components/schemas/TeamUser'
...
    AlertsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - alerts
              properties:
                alerts:
                  type: array
                  items:
                    $ref: '#/components/schemas/Alert'
...
    TrendingTopicsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - topics
              properties:
                topics:
                  type: array
                  items:
                    $ref: '#/components/schemas/TrendingTopic'